requirements.txt
cffi==1.17.1
cryptography==43.0.1
html2text==2024.2.26
lxml==5.3.0
numpy==2.1.1
oracledb==2.4.1
pandas==2.2.3
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
